requirements.txt
# Core dependencies
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
datasets>=2.14.0
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.1.0
omegaconf>=2.3.0
hydra-core>=1.3.0
pyyaml>=6.0.0
tqdm>=4.65.0
rich>=13.6.0

# API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
httpx>=0.25.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9

# Model dependencies
einops>=0.7.0
flash-attn>=2.3.0
xformers>=0.0.22
triton>=2.1.0
apex>=0.1
bitsandbytes>=0.41.0

# Agent dependencies
openai>=1.3.0
langchain>=0.0.340
tiktoken>=0.5.0

# Evaluation dependencies
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.0
nltk>=3.8.0

# Tracking and monitoring
wandb>=0.16.0
mlflow>=2.8.0
tensorboard>=2.15.0
prometheus-client>=0.19.0

# Data processing
h5py>=3.10.0
zarr>=2.16.0
pyarrow>=14.0.0
faiss-cpu>=1.7.4

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-xdist>=3.4.0
pytest-benchmark>=4.0.0
hypothesis>=6.92.0

# Code quality
mypy>=1.6.0
ruff>=0.1.0
black>=23.11.0
isort>=5.12.0
pre-commit>=3.5.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=2.0.0
myst-parser>=2.0.0
autodoc-pydantic>=2.0.0

# Optimization
optuna>=3.4.0
scikit-optimize>=0.9.0
hyperopt>=0.2.7
ray[tune]>=2.8.0

# Utilities
click>=8.1.0
python-dotenv>=1.0.0
tenacity>=8.2.0
joblib>=1.3.0
filelock>=3.13.0